controller delete post fix, cart page fix bugs

+++ vylka/Controllers/Admin/UserAdminController.cs
@@ -2,12 +2,11 @@
 
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using vylka.Areas.Entity;
 using vylka.Areas.Identity.Data;
 
 namespace vylka.Controllers.Admin
 {
-    [Authorize(Roles = "Адмін")]
+    /*[Authorize(Roles = "Адмін")]*/
     public class UserAdminController : Controller
     {
         private readonly vylkaContext _context;
@@ -41,21 +40,14 @@ namespace vylka.Controllers.Admin
         {
             var obj = _context.Users.Find(id);
             var cart = _context.Cart.OrderBy(o => o.Id).LastOrDefault(u => u.CartUserId.Id == id);
-            var items = _context.ShippingDetail.Where(u => u.UserId == id);
-            
+
             if (obj == null || cart == null)
             {
                 return NotFound();
             }
 
-            if (items != null)
-            {
-                foreach (var item in _context.ShippingDetail.Where(u => u.UserId == id))
-                {
-                    _context.ShippingDetail.Remove(item);
-                }
-            }
-            
+            foreach (var item in _context.ShippingDetail.Where(u => u.UserId == id)) _context.ShippingDetail.Remove(item);
+
             _context.Cart.Remove(cart);
             _context.Users.Remove(obj);
             _context.SaveChanges();
